Bestiary model reorganization, added AwakenCost model
