+++ herders/templates/herders/profile/teams/team_detail.html
@@ -105,15 +105,17 @@
                         {% endif %}
                     </td>
                     <td class="monster-skill-effects">
-                        {% for effect in team.leader.monster.all_skill_effects %}
-                            {% if effect.icon_filename %}
-                            <img src="{{ img_url_prefix }}buffs/{{ effect.icon_filename }}"
-                                data-toggle="popover" data-trigger="hover" data-placement="top" data-container="body" title="{{ effect.name }}" data-content="{{ effect.description }}"/>
-                            {% else %}
-                                <span class="skill-effect {% if effect.is_buff %}skill-effect-buff{% else %}skill-effect-debuff{% endif %}" data-toggle="popover" data-trigger="hover" data-placement="top" data-container="body" title="{{ effect.name }}" data-content="{{ effect.description }}">
-                                    <span>{{ effect.name }}</span>
-                                </span>
-                            {% endif %}
+                        {% for skill in team.leader.monster.skills.all %}
+                            {% for effect in skill.skill_effect.all %}
+                                {% if effect.icon_filename %}
+                                <img src="{{ img_url_prefix }}buffs/{{ effect.icon_filename }}"
+                                    data-toggle="popover" data-trigger="hover" data-placement="top" data-container="body" title="{{ effect.name }}" data-content="{{ effect.description }}"/>
+                                {% else %}
+                                    <span class="skill-effect {% if effect.is_buff %}skill-effect-buff{% else %}skill-effect-debuff{% endif %}" data-toggle="popover" data-trigger="hover" data-placement="top" data-container="body" title="{{ effect.name }}" data-content="{{ effect.description }}">
+                                        <span>{{ effect.name }}</span>
+                                    </span>
+                                {% endif %}
+                            {% endfor %}
                         {% endfor %}
                     </td>
 
@@ -167,15 +169,17 @@
                             {% endif %}
                         </td>
                         <td class="monster-skill-effects">
-                            {% for effect in instance.monster.all_skill_effects %}
-                                {% if effect.icon_filename %}
-                                    <img src="{{ img_url_prefix }}buffs/{{ effect.icon_filename }}"
-                                            data-toggle="popover" data-trigger="hover" data-placement="top" data-container="body" title="{{ effect.name }}" data-content="{{ effect.description }}"/>
-                                {% else %}
-                                    <span class="skill-effect {% if effect.is_buff %}skill-effect-buff{% else %}skill-effect-debuff{% endif %}" data-toggle="popover" data-trigger="hover" data-placement="top" data-container="body" title="{{ effect.name }}" data-content="{{ effect.description }}">
-                                        <span>{{ effect.name }}</span>
-                                    </span>
-                                {% endif %}
+                            {% for skill in instance.monster.skills.all %}
+                                {% for effect in skill.skill_effect.all %}
+                                    {% if effect.icon_filename %}
+                                        <img src="{{ img_url_prefix }}buffs/{{ effect.icon_filename }}"
+                                                data-toggle="popover" data-trigger="hover" data-placement="top" data-container="body" title="{{ effect.name }}" data-content="{{ effect.description }}"/>
+                                    {% else %}
+                                        <span class="skill-effect {% if effect.is_buff %}skill-effect-buff{% else %}skill-effect-debuff{% endif %}" data-toggle="popover" data-trigger="hover" data-placement="top" data-container="body" title="{{ effect.name }}" data-content="{{ effect.description }}">
+                                            <span>{{ effect.name }}</span>
+                                        </span>
+                                    {% endif %}
+                                {% endfor %}
                             {% endfor %}
                         </td>
                     </tr>
